minor reorg: move parse_val mod out of engine mod; expose it publicly, make engine code shorter
