+++ src/engine.rs
@@ -38,8 +38,6 @@ use std::hash::{Hash,Hasher};
 use std::collections::hash_map::DefaultHasher;
 use std::mem::replace;
 use std::mem::transmute;
-//use std::num::Zero;
-//use std::ops::Add;
 use std::rc::Rc;
 use std::fmt::Write;
 
@@ -83,9 +81,10 @@ fn my_hash<T>(obj: T) -> u64
 
 pub mod reflect {
   pub use reflect::*;
-  use std::fmt::{Debug,Write};
+  pub use parse_val;
+
+  use std::fmt::{Write};
   use super::{TraceSt,TRACES,GLOBALS,Engine};
-  use super::parse_val;
   use adapton::engine::Name;
 
   /// See doc for `write_name`. Returns this output as a string.
@@ -123,31 +122,6 @@ pub mod reflect {
     write_path(w, &l.path);
     write!(w, "_").unwrap(); // Underscores are valid in CSS class names
     write_name(w, &l.name);
-  }
-
-  /// _Rust data and articulation reflection_: Transform any(*) Rust
-  /// data that derives `Debug` into a reflected `Val`.  
-  /// 
-  /// This parsing logic handles user-defined `struct` and `enum`
-  /// types, tuples and vectors.  It recognizes the `Debug` output of
-  /// these structures and parses them into trees of type
-  /// `Val`. Importantly, it recognizes articulations in this `Debug`
-  /// output and parses those into reflected locations (of type `Loc`)
-  /// and articulations (of type `Art`).  The reflected `DCG` maps
-  /// reflected articulations (whose reflected locations are of type
-  /// `Loc`) to reflected nodes that contain more reflected values.
-  /// 
-  /// (*) Note: Though this parsing logic handles vectors, tuples and
-  /// user-defined data types, this parsing logic is probably
-  /// incomplete for all of Rust's standard collections. (It does not
-  /// yet handle `HashMap<_,_>` debug output, or anything else of this
-  /// complexity as of yet).
-  pub fn reflect_val <V:Debug> (v:&V) -> Val {
-    let s = format!("{:?}", v);
-    //println!("reflect_val({:?})", v);
-    let toks = parse_val::lex(s.into_bytes());
-    //println!("toks = {:?}", toks);
-    parse_val::parse_val(toks)
   }
   
   /// Reflect the DCG's internal structure now.  Does not reflect any
@@ -600,6 +574,7 @@ enum Node<Res> {
 }
 impl<X:Debug> reflect::Reflect<reflect::Node> for Node<X> {
   fn reflect(&self) -> reflect::Node {
+    use parse_val::parse_val;
     match *self {
       Node::Comp(ref n) => {
         reflect::Node::Comp(
@@ -608,7 +583,7 @@ impl<X:Debug> reflect::Reflect<reflect::Node> for Node<X> {
             succs:n.succs.reflect(),
             prog_pt:n.producer.prog_pt().clone(),
             value:match n.res { 
-              Some(ref v) => Some( reflect::reflect_val(v) ),
+              Some(ref v) => Some( parse_val(v) ),
               None => None
             }
           })
@@ -616,14 +591,14 @@ impl<X:Debug> reflect::Reflect<reflect::Node> for Node<X> {
       Node::Pure(ref n) => {
         reflect::Node::Pure(
           reflect::PureNode {
-            value:reflect::reflect_val( &n.val ),
+            value:parse_val( &n.val ),
           })
       },
       Node::Mut(ref n) => {
         reflect::Node::Ref(
           reflect::RefNode {
             preds:n.preds.reflect(),
-            value:reflect::reflect_val( &n.val ),
+            value:parse_val( &n.val ),
           })        
       },
       Node::Unused => panic!(""),
@@ -687,61 +662,6 @@ pub enum ArtIdChoice {
   /// Identifies an `Art` based on a programmer-chosen name.
   Nominal(Name),
 }
-
-// /// *Engine Counts*: Metrics that reflect the time and space costs of the engine.
-// #[derive(Debug,Hash,PartialEq,Eq,Clone)]
-// pub struct Cnt {
-//   /// Number of DCG nodes created
-//   pub create : usize, // Add trait performs sum
-//   /// Number of DCG nodes evaluated
-//   pub eval   : usize, // Add trait performs sum
-//   /// Number of DCG nodes marked as dirty
-//   pub dirty  : usize, // Add trait performs sum
-//   /// Number of DCG nodes reverted from dirty to clean
-//   pub clean  : usize, // Add trait performs sum
-//   /// Maximum height of the DCG node stack.  This stack is pushed when DCG nodes are evaluated, and popped when they complete.
-//   pub stack  : usize, // Add trait performs max
-// }
-
-// impl Add for Cnt {
-//   type Output=Cnt;
-//   fn add(self, rhs: Self) -> Self::Output {
-//     Cnt {
-//       create : self.create + rhs.create,
-//       eval   : self.eval + rhs.eval,
-//       dirty  : self.dirty + rhs.dirty,
-//       clean  : self.clean + rhs.clean,
-//       stack  : if self.stack > rhs.stack { self.stack } else { rhs.stack }
-//     }
-//   }
-// }
-
-// impl<'a> Add for &'a Cnt {
-//   type Output=Cnt;
-//   fn add(self, rhs: Self) -> Self::Output {
-//     Cnt {
-//       create : self.create + rhs.create,
-//       eval   : self.eval + rhs.eval,
-//       dirty  : self.dirty + rhs.dirty,
-//       clean  : self.clean + rhs.clean,
-//       stack  : if self.stack > rhs.stack { self.stack } else { rhs.stack }
-//     }
-//   }
-// }
-
-// impl Zero for Cnt {
-//   fn zero() -> Self {
-//     Cnt {
-//       create : 0 as usize,
-//       eval   : 0 as usize,
-//       dirty  : 0 as usize,
-//       clean  : 0 as usize,
-//       stack  : 0 as usize,
-//     }
-//   }
-// }
-
-
 
 // Produce a value of type Res.
 trait Producer<Res> : Debug {
@@ -1138,8 +1058,6 @@ impl <Res:'static+Sized+Debug+PartialEq+Eq+Clone+Hash>
   }
 
   fn clean(self:&Self, g:&RefCell<DCG>, loc:&Rc<Loc>) -> DCGRes {
-    //let stackLen = mut_dcg_of_globals.stack.len() ;
-    ////debug!("{} change_prop begin: {:?}", engineMsg!(st), loc);
     let res_succs = { // Handle cases where there is no internal computation to re-compute:
       let st = &mut *g.borrow_mut();
       let node : &mut Node<Res> = res_node_of_loc(st, loc) ;
@@ -1150,11 +1068,9 @@ impl <Res:'static+Sized+Debug+PartialEq+Eq+Clone+Hash>
             None => None
           }},
         Node::Pure(_) => {
-          ////debug!("{} change_prop early end: {:?} is Pure(_)", engineMsg(Some(stackLen)), loc);
           return DCGRes{changed:false}
         },
         Node::Mut(ref nd) => {
-          ////debug!("{} change_prop early end: {:?} is Mut(_)", engineMsg(Some(stackLen)), loc);
           return DCGRes{changed:nd.val != self.res}
         },
         _ => panic!("undefined")
@@ -1342,8 +1258,6 @@ trait Adapton : Debug+PartialEq+Eq+Hash+Clone {
   /// Enters a special "namespace" where all name uses are ignored; instead, Adapton uses structural identity.
   fn structural<T,F> (g: &RefCell<DCG>, body:F) -> T where F:FnOnce() -> T;
   
- //fn cnt<Res,F> (g: &RefCell<DCG>, body:F) -> (Res, Cnt) where F:FnOnce() -> Res;
-  
   /// Creates immutable, eager articulation.
   fn put<T:Eq+Debug+Clone> (self:&mut Self, T) -> AbsArt<T,Self::Loc> ;
   
@@ -1395,7 +1309,6 @@ impl Adapton for DCG {
       table : table,
       stack : stack,
       path  : path,
-//      cnt   : Cnt::zero (),
       dcg_count : 0,
       dcg_hash : 0, // XXX This makes assumptions about hashing implementation
     }
@@ -1428,25 +1341,6 @@ impl Adapton for DCG {
     g.borrow_mut().path = saved ;
     x
   }
-
-  // fn cnt<Res,F> (g: &RefCell<DCG>, body:F) -> (Res,Cnt)
-  //   where F:FnOnce() -> Res
-  // {    
-  //   let c : Cnt = {
-  //     let st = &mut *g.borrow_mut();
-  //     let c = st.cnt.clone() ;
-  //     st.cnt = Zero::zero();
-  //     c
-  //   };
-  //   let x = body() ;
-  //   let d : Cnt = {
-  //     let st = &mut *g.borrow_mut();      
-  //     let d : Cnt = st.cnt.clone() ;
-  //     st.cnt = c + d.clone();
-  //     d
-  //   };
-  //   (x, d)
-  // }
 
   fn put<T:Eq> (self:&mut DCG, x:T) -> AbsArt<T,Self::Loc> { AbsArt::Rc(Rc::new(x)) }
 
@@ -1522,7 +1416,6 @@ impl Adapton for DCG {
             preds:Vec::new(),
             val:val.clone(),
           })} ;
-        //self.cnt.create += 1;                    
         self.table.insert(loc.clone(), Box::new(node));
       } ;
       if ! is_pure { match self.stack.last_mut() { 
@@ -1711,7 +1604,6 @@ impl Adapton for DCG {
             producer:Box::new(producer),
             res:None,
           } ;
-          //self.cnt.create += 1;
           self.table.insert(loc.clone(), Box::new(Node::Comp(node)));
           wf::check_dcg(self);
           AbsArt::Loc(loc)
@@ -2067,10 +1959,12 @@ pub fn name_pair (n1:Name, n2:Name) -> Name {
   Name{ hash:h, symbol:Rc::new(p) }
 }
 
+/// Create a name from a hash value.
+///
 /// Do not call this function directly; we introduced it in order to
 /// get reflection to type-check.  We should think of ways to avoid
 /// using this in the future.
-fn name_of_hash64(h:u64) -> Name {
+pub fn name_of_hash64(h:u64) -> Name {
   // TODO: Get rid of need for Rc here; 
   // Rc should be optional in names?
   Name{ hash:h, symbol:Rc::new(NameSym::Hash64) }
@@ -2335,17 +2229,6 @@ pub mod manage {
   pub fn init_engine (engine: Engine) -> Engine {
     use_engine(engine)
   }  
-  
-  // /// Counts various engine cost metrics, returning a product of sums (`Cnt`)
-  // pub fn cnt<Res,F> (body:F) -> (Res, Cnt)
-  //   where F:FnOnce() -> Res {
-  //   GLOBALS.with(|g| {
-  //     match g.borrow().engine {
-  //       Engine::DCG(ref dcg) => <DCG as Adapton>::cnt(dcg,body),
-  //       Engine::Naive => ((body)(), Cnt::zero())
-  //     }
-  //   })
-  // }
 
   /// True iff the current engine is `Naive`
   pub fn engine_is_naive () -> bool {
@@ -2474,13 +2357,11 @@ mod wf {
     let mut writer = BufWriter::new(file);
     writeln!(&mut writer, "digraph {{\n").unwrap();
     writeln!(&mut writer, "ordering=out;").unwrap();
-    //let mut frame_num = 0;
     for frame in st.stack.iter() {
       writeln!(&mut writer, "\"{:?}\" [color=blue,penwidth=10];", frame.loc).unwrap();
       for succ in frame.succs.iter() {
         writeln!(&mut writer, "\"{:?}\" -> \"{:?}\" [color=blue,weight=10,penwidth=10];", &frame.loc, &succ.0.loc).unwrap();
       }
-      //frame_num += 1;
     };
     for (loc, node) in &st.table {
       if ! node.succs_def () {
@@ -2538,429 +2419,5 @@ mod wf {
         assert!( succ.dirty ); // The edge is clean.
       }
     }
-  }
-}
-
-/// Parses the output of Rust `Debug` strings into reflected `Val`
-/// values.  We use this parse as a non-intrusive mechanism for
-/// building the values in the reflected DCG, which consists of
-/// crawing user-defined data structures, and following their
-/// articulations. We use the values' `Debug` strings to do this
-/// traversal.
-mod parse_val { 
-  use adapton::engine::reflect::{Loc,Path,Val,ArtContent,Const};
-  use adapton::engine::{Name, name_of_str, name_of_string};
-
-  /// _Balanced tokens_: Tokens that must be balanced with a left and
-  /// right instance and well-nested balanced tokens between them.
-  #[derive(Debug, Eq, PartialEq)]
-  pub enum BalTok {
-    Paren, 
-    Bracket, 
-    Brace
-  }
-
-  // #[derive(Debug, Eq, PartialEq)]
-  // pub enum Const {
-  //   Nat(usize),
-  //   String(String),
-  // }
-
-  /// _Tokens_: The unit of input for parsing Rust `Debug` strings
-  /// into reflected `Val` values.
-  #[derive(Debug, Eq, PartialEq)]
-  pub enum Tok {
-    /// Left (and right) balanced tokens
-    Left(BalTok), 
-    /// Right (and left) balanced tokens
-    Right(BalTok),
-    /// Constant values that can immediately be injected into reflected `Val` type
-    Const(Const),
-    /// Identifers name fields of structs and constructors (of enums and structs)
-    Ident(String),
-    /// Colons separate field names from field values in structs.  Co
-    Colon,
-    /// Commas separate arguments to a constructor; for struct constructors, they separate fields
-    Comma, 
-  }
-
-  /// Tokenize the characters of input into lexical tokens of type `Tok`
-  pub fn lex (mut chars: Vec<u8>) -> Vec<Tok> {
-    let mut toks = vec![];
-    chars.reverse(); // TODO rewrite to avoid this
-    loop {
-      match chars.pop() {
-        None => return toks,
-        Some(c) => {
-          let c : char = c as char ;
-          if      c == ' ' { continue }
-          else if c == ':' { toks.push(Tok::Colon); continue }
-          else if c == ',' { toks.push(Tok::Comma); continue }
-          else if c == '{' { toks.push(Tok::Left (BalTok::Brace));   continue }
-          else if c == '[' { toks.push(Tok::Left (BalTok::Bracket)); continue }
-          else if c == '(' { toks.push(Tok::Left (BalTok::Paren));   continue }
-          else if c == '}' { toks.push(Tok::Right(BalTok::Brace));   continue }
-          else if c == ']' { toks.push(Tok::Right(BalTok::Bracket)); continue }
-          else if c == ')' { toks.push(Tok::Right(BalTok::Paren));   continue }
-          else if c == '"' {
-            let mut string_chars = vec![];
-            loop {
-              match chars.pop() {
-                None => break,
-                Some(c) => {
-                  let c : char = c as char ;
-                  if c == '"' { break } else { 
-                    string_chars.push(c);
-                    continue 
-                  }
-                }
-              }
-            };
-            toks.push(Tok::Const(Const::String( 
-              string_chars.into_iter().collect() 
-            )));
-            continue
-          }
-          else if c == '-' || (c >= '0' && c <= '9') {
-            let mut digs = vec![c];
-            loop {
-              match chars.pop() {
-                None    => break,
-                Some(c) => { 
-                  let c : char = c as char ;
-                  if c >= '0' && c <= '9' { 
-                    digs.push(c); 
-                    continue 
-                  } else { 
-                    chars.push(c as u8); 
-                    break 
-                  }
-                }
-              }
-            };
-            if c == '-' {
-              let s : String = digs.into_iter().collect();
-              toks.push(Tok::Const(Const::Num( 
-                isize::from_str_radix(s.as_str(), 10).unwrap()
-              )));              
-            } else {
-              let s : String = digs.into_iter().collect();
-              toks.push(Tok::Const(Const::Nat( 
-                usize::from_str_radix(s.as_str(), 10).unwrap()
-              )));
-            }
-            continue
-          }
-          else if (c >= 'a' && c <= 'z') ||
-            (c >= 'A' && c <= 'Z') ||
-            (c == '_') 
-          {
-            let mut ident = vec![c];
-            loop {
-              match chars.pop() {
-                None    => break,
-                Some(c) => { 
-                  let c : char = c as char ;
-                  if (c >= 'a' && c <= 'z') ||
-                    (c >= 'A' && c <= 'Z') ||
-                    (c >= '0' && c <= '9') ||
-                    (c == '_')  
-                  {
-                    ident.push(c); 
-                    continue 
-                  } else { 
-                    chars.push(c as u8); 
-                    break 
-                  }
-                }
-              }
-            };
-            toks.push(Tok::Ident( ident.into_iter().collect() ));
-            continue           
-          }
-        }
-      }
-    } ;
-  }
-
-  /// Parse a sequence of fields (appending to `fields`) until right
-  /// balanced token `bal`.  Return fields and remaining tokens.
-  pub fn parse_fields (mut toks:Vec<Tok>, mut fields:Vec<(Name, Val)>, bal:Tok) -> (Vec<(Name, Val)>, Vec<Tok>) {
-    match toks.pop() {
-      None => panic!("parse_vals: expected more vals, or end of sequence; but no more tokens"),
-      Some(t) => {
-        if t == bal { (fields, toks) } 
-        else if t == Tok::Comma { 
-          return parse_fields(toks, fields, bal)
-        } else {
-          match t {
-            Tok::Ident(i) => {
-              let toks = expect_tok(toks, Tok::Colon);
-              let (v, toks) = parse_val_rec(toks);
-              fields.push((name_of_string(i), v));
-              return parse_fields(toks, fields, bal)
-            }
-            t => {
-              panic!("parse_fields: expected identifier, but found {:?}", t)
-            }
-          }
-        }
-      }
-    }
-  }
-
-  /// Parse a sequence of values (appending to `vals`) until right
-  /// balanced token `bal`.  Return fields and remaining tokens.
-  pub fn parse_vals (mut toks:Vec<Tok>, mut vals:Vec<Val>, bal:Tok) -> (Vec<Val>, Vec<Tok>) {
-    match toks.pop() {
-      None => panic!("parse_vals: expected more vals, or end of sequence; but no more tokens"),
-      Some(t) => {
-        if t == bal { (vals, toks) } 
-        else if t == Tok::Comma { 
-          return parse_vals(toks, vals, bal)
-        } 
-        else {
-          toks.push(t);
-          let (v, toks) = parse_val_rec(toks);
-          vals.push(v);
-          return parse_vals(toks, vals, bal)
-        }
-      }
-    }
-  }
-
-  /// Expect next token to be `tok` and panic otherwise.
-  pub fn expect_tok (mut toks: Vec<Tok>, tok:Tok) -> Vec<Tok> {
-    match toks.pop() {
-      None => panic!("expected token `{:?}`, but, no more tokens", tok),
-      Some(t) => {
-        if t == tok { toks } 
-        else { panic!("expected token `{:?}`, but instead found token `{:?}`", tok, t) }
-      }
-    }
-  }
-
-  pub fn parse_val(mut toks:Vec<Tok>) -> Val {
-    toks.reverse();
-    let (v, toks) = parse_val_rec(toks);
-    assert!(toks.len() == 0);
-    v
-  }
-
-  pub fn path_of_val ( p:&Val ) -> Path {
-    match *p {
-      Val::Vec( ref vs ) => vs.iter().map( name_of_val ).collect(),
-      _ => panic!("expected a vector of values representing names"),
-    }
-  }
-
-  pub fn name_of_val ( n:&Val ) -> Name {
-    use engine::*;
-
-    match *n {
-      Val::Constr( ref cons_name, ref cons_args ) => {
-        if *cons_name == name_of_str("Unit") {
-          name_unit()
-        }
-        else if *cons_name == name_of_str("Hash64") {
-          name_of_hash64( 0 ) // TODO/XXX
-        }
-        else if *cons_name == name_of_str("String") {
-          name_of_string( match cons_args[0] {
-            Val::Const( Const::String( ref s ) ) => s.clone(),
-            _ => panic!("expected a String"),
-          })
-        }
-        else if *cons_name == name_of_str("Usize") {
-          name_of_usize( match cons_args[0] {
-            Val::Const( Const::Nat( ref n ) ) => n.clone(),
-            _ => panic!("expected a Nat"),
-          })
-        }
-        else if *cons_name == name_of_str("Isize") {
-          panic!("")
-        }
-        else if *cons_name == name_of_str("Pair") {
-          let n1 = name_of_val( & cons_args[0] );
-          let n2 = name_of_val( & cons_args[1] );
-          name_pair(n1, n2)
-        }
-        else if *cons_name == name_of_str("ForkL") {
-          let n = name_of_val( & cons_args[0] );
-          name_fork(n).0
-        }
-        else if *cons_name == name_of_str("ForkR") {
-          let n = name_of_val( & cons_args[0] );
-          name_fork(n).1
-        }
-        else {
-          unreachable!()
-        }        
-      },
-      Val::Name(ref n) => n.clone(),
-      _ => panic!("expected a constructor for a NameSym")
-    }
-  }
-
-  pub fn name_option_of_val ( n:&Val ) -> Option<Name> {
-    use engine::*;
-    
-    match *n {
-      Val::Constr( ref cons_name, ref cons_args ) => {
-        if *cons_name == name_of_str("Unit") {
-          Some(name_unit())
-        }
-        else if *cons_name == name_of_str("Hash64") {
-          Some(name_of_hash64( 0 )) // XXX \ TODO -- We are actually missing the hash here!
-        }
-        else if *cons_name == name_of_str("String") {
-          if cons_args.len() < 1 { None } else {
-            match cons_args[0] {            
-              Val::Const( Const::String( ref s ) ) => 
-                Some(name_of_string( s.clone() )),
-              _ => None,
-            }}
-        }
-        else if *cons_name == name_of_str("Usize") {
-          if cons_args.len() < 1 { None } else {
-            match cons_args[0] {          
-              Val::Const( Const::Nat( ref n ) ) => Some( name_of_usize( n.clone() ) ),
-              _ => None,
-            }}
-        }
-        else if *cons_name == name_of_str("Isize") {
-          panic!("TODO")
-        }
-        else if *cons_name == name_of_str("Pair") {          
-          if cons_args.len() < 2 { None } else {
-            let n1 = name_option_of_val( & cons_args[0] );
-            let n2 = name_option_of_val( & cons_args[1] );
-            if cons_args.len() < 2 { None } else {
-              match (n1,n2) {
-                (Some(n1),Some(n2)) => Some(name_pair(n1, n2)),
-                (_, _) => None,
-              }}}
-        }
-        else if *cons_name == name_of_str("ForkL") {
-          if cons_args.len() < 1 { None } else {
-            let n = name_option_of_val( & cons_args[0] );
-            match n {
-              None => None,
-              Some(n) => Some(name_fork(n).0)
-            }}
-        }
-        else if *cons_name == name_of_str("ForkR") {
-          if cons_args.len() < 1 { None } else {
-            let n = name_option_of_val( & cons_args[0] );
-            match n {
-              None => None,
-              Some(n) => Some(name_fork(n).1)
-            }}
-        }
-        else { None }
-      },
-      Val::Name(ref n) => Some(n.clone()),
-      _ => None,
-    }
-  }
- 
-
-  /// Attempts to parse a reflected value into an `Art` value case,
-  /// which consists of parsing a location represented as a `Val`
-  /// structure into a `Loc` represented as a Rust data type (in the
-  /// `reflect` module).  If it fails to parse a value into an art, it
-  /// returns None.
-  pub fn parse_art_val ( i:&String, fields:&Vec<(Name, Val)> ) -> Option<Val> {
-    if i == "Art" && fields.len() == 1 { 
-      match fields[0] { 
-        (ref nf, ref vf) =>
-          if *nf == name_of_str("art") {
-            // OK: it's a struct called Art with exactly one field
-            // called art.  We are going to parse this into an Art.
-
-            match *vf { 
-              Val::Struct( ref j, ref ws ) => 
-                if *j == name_of_str("Loc") 
-                && ws.len() == 2
-                && ws[0].0 == name_of_str("path") 
-                && ws[1].0 == name_of_str("id")
-              {
-                // Now we are confident that the rest ought to parse.
-                // Any further parse errors are panics.
-                let path = path_of_val( & ws[0].1 );
-                let name = name_of_val( & ws[1].1 );
-                Some( Val::Art(Loc{path:path, name:name}, ArtContent::Unknown) )
-              } 
-
-
-              else { 
-                None 
-              },
-              _ => None,
-            }
-          } else { None }
-      }} else { None }           
-  }
-
-  /// Parse a value from the tokens `toks` and return it.  Panic if the next tokens do not parse into value.
-  pub fn parse_val_rec (mut toks:Vec<Tok>) -> (Val, Vec<Tok>) {
-    //println!("{:?}", toks);
-    let (v, toks) = match toks.pop() {
-      None => panic!("expected value; but, no more tokens"),
-      Some(Tok::Right(r)) => panic!("expected value, but found {:?} instead", Tok::Right(r)),
-      Some(Tok::Comma) => panic!("expected value, but found Comma instead"),
-      Some(Tok::Colon) => panic!("expected value, but found Colon instead"),
-      Some(Tok::Left(BalTok::Bracket)) => {
-        // Parse a vector: Begins with '[', then a list of comma-separated values, then ']'.
-        let (vs, toks) = parse_vals(toks, vec![], Tok::Right(BalTok::Bracket));
-        (Val::Vec(vs), toks)
-      },
-      Some(Tok::Left(BalTok::Paren)) => {
-        // Parse a tuple: Begins with '(', then a list of comma-separated values, then ')'.
-        let (vs, toks) = parse_vals(toks, vec![], Tok::Right(BalTok::Paren));
-        (Val::Tuple(vs), toks)
-      },
-      Some(Tok::Left(l)) => panic!("expected value, but found {:?} instead", Tok::Left(l)),     
-      Some(Tok::Ident(i)) => {
-        match toks.pop() {
-          None => {
-            // Constructors with no arguments (e.g., Nil)
-            (Val::Constr(name_of_string(i), vec![]), toks)
-          }
-          Some(Tok::Left(BalTok::Brace)) => {
-            //println!("parsing struct: {:?}", i);
-            let (fields, toks) = parse_fields(toks, vec![], Tok::Right(BalTok::Brace));
-            let art_op = parse_art_val(&i, &fields);
-            let v = match art_op {
-              Some(a) => a,
-              None => Val::Struct(name_of_string(i.clone()), fields.clone())
-            };    
-            (v, toks)
-          }
-          Some(Tok::Left(BalTok::Paren)) => {
-            //println!("parsing constructor: {:?}", i);
-            let (vs, toks) = parse_vals(toks, vec![], Tok::Right(BalTok::Paren));
-            let v = Val::Constr(name_of_string(i), vs);
-            match name_option_of_val(&v) {
-              Some(n) => (Val::Name(n), toks),
-              None => (v, toks)
-            }
-          },
-          Some(Tok::Comma) => {
-            toks.push(Tok::Comma);
-            (Val::Constr(name_of_string(i), vec![]), toks)
-          },
-          Some(Tok::Right(baltok)) => {
-            toks.push(Tok::Right(baltok));
-            (Val::Constr(name_of_string(i), vec![]), toks)
-          },
-          Some(t) => {
-            panic!("expected left balanced token, or comma, but instead found token {:?}", t)
-          }}},
-        Some(Tok::Const(Const::Nat(n)))    => (Val::Const(Const::Nat(n)), toks),
-        Some(Tok::Const(Const::Num(n)))    => (Val::Const(Const::Num(n)), toks),
-        Some(Tok::Const(Const::String(s))) => (Val::Const(Const::String(s)), toks)
-    };
-    (v,toks)
   }
 }
